Ajustes nas rotas de deletar

+++ src/modules/accounts/UseCases/DeleteUser/DeleteUserUserCase.ts
@@ -1,3 +1,4 @@
+import { AppErros } from "../../../../errors/AppErros";
 import { IDeleteUserUserCase } from "../../Protocols/DeleteUser/IDeleteUserUserCase";
 import { IUserRepository } from "../../Protocols/IUserRepository";
 
@@ -5,6 +6,10 @@ import { IUserRepository } from "../../Protocols/IUserRepository";
 export class DeleteUserUserCase implements IDeleteUserUserCase{
     constructor(private userRepository: IUserRepository){}
     async execute(id:string): Promise<void>{
+        const user=await this.userRepository.findById(id)
+        if(!user){
+            throw new AppErros('User not exists')
+        }
         await this.userRepository.delete(id)
     }
 }
